Refactors `SecureSumFactory` to make sure compatible bound dtypes work as expected.
For example, previously, a tf.float64 `value_type` with specified python float bounds would raise TypeError due to implicit conversion to tf.float32.

This change also makes it explicit that a structure `value_type` must have all dtypes identical. This might lead to numerical surprises, so better to disallow it until we would really need it. Note this was not supported previously, now a clear error message is raised. This also implements a general type checking utility.

Finally, enables the bounds to be specified as numpy scalars, instead of previous np ndarrays containing a scalar.

PiperOrigin-RevId: 425375469