Add an example C++ implementation of Python DataBackend
Currently, both C++ and Python implementations of the DataBackend exist. In order to allow gradual migration of users from the Python runtime to the C++ runtime, it is suggested that users develop agains the C++ DataBackend interface. They can then use their C++ DataBackend from Python as shown here prior to making a full transition to the C++ runtime.

PiperOrigin-RevId: 425634355