Preserve more names during unique-ification
Name unique-ification previously replaced every single name binding. This resulted
in computations that were harder to read and understand and which lost more of the context from their original binding.

This CL changes name unique-ification to instead only rename bindings which are either duplicates of another binding, or fall back to renaming all bindings in the case that an explicit name_generator has been provided (suggesting there is some unknown outer set of bindings with which the computation must not conflict).

PiperOrigin-RevId: 430778496