Adds ingest-corofunc case to argument packing, allowing for simpler implementations of async def ingest.
Previous attempts to implement async def ingest functions were difficult to compose with the behavior of pack_args when repackaging as a structure was necessary. This change introduces an invariant that an async def ingest can always expect to be called with either a raw value coming in from user-land, or a coroutine which returns a structure of (awaited and) ingested values when awaited.

It is worth noting that this behavior could additionally be implemented in the body of async def ingest functions; any structure passed to ingest could be traversed (at the top level) and have its elements awaited (if they are coroutines). This would avoid the nasty union in return types for `pack_args_into_struct`. I preferred this route due to some experience programming async contexts, and introducing them to new users--this invariant, I think, might make life easier on future context implementers (when their contexts expect to be composed with `ConcreteComputation`).

PiperOrigin-RevId: 437153366