Changes service status returned when an executor is not found to FAILED_PRECONDITION.
Adds FAILED_PRECONDITION to retryable errors.

Re-adds explicit check/raise in Python, raising RuntimeError if no executor is found. This represents the previous behavior.

Thinking about our missing coverage here: we believe we have testing for the case that a worker goes down between invocations, but no explicit coverage for the system cleaning itself up appropriately when a worker fails mid-invocation. This should trigger a retry (at the appropriate level of granularity, which today is the entire invocation). This CL reintroduces this behavior.

PiperOrigin-RevId: 438106121