Finish support for federated_select on the C++ runtime
The primary focus of this CL is implementing the args_into_sequence intrinsic on the TFF C++ runtime's TensorFlow Executor. This is the final step necessary for executing federated_select computations on the C++ runtime.

This CL also adds an integration test ensuring that federated_select is able to run on all execution backends.

Note that this CL won't be submittable immediately, as unfortunately the use of DatasetFromTensorSlices in the TensorFlowExecutor will break the OSS Python->Native bindings until it's possible to use both the TF C++ APIs and TF Python bindings within the same process. This issue is tracked in b/209816646.

PiperOrigin-RevId: 442939260