Add support for direct modification and reading from the client update checkpoint
This can be used to perform custom appending to the client checkpoint using the AppendSlices op, and then read the results back in during the execution of `federated_aggregate`'s `accumulate` operation. Users doing so must ensure that they apply a `control_deps` on any append operations to ensure these operations run before the simple aggregation variables (those going through the more typical federated_aggregate pathway) are written out.

Note that use of this is not yet broadly encouraged: this pathway is not supported by TFF simulations, and is generally in contradiction to the functional style TFF is designed to support. Later changes may resolve these contradictions and render this API more broadly usable, but this is not yet planned.

PiperOrigin-RevId: 444934493