Move the `tf.errors.PermissionDeniedError` exception handling so that it captures failures to close the file handle.
When the file handle is closed as a result of an error, this change correctly returns the `FileReleaseManagerPermissionDeniedError`, "Could not append a value to the file '...'. It is possible that this file is compressed or encoded. Please use write mode instead of append mode to release values to this file using a `tff.program.CSVFileReleaseManager`."

PiperOrigin-RevId: 446203127