Move asserts from `tff.test.TestCase` to `tff.test.*` as functions.
Internally it's easier to use the functions than the class, because the functions can be broken up and owned by the package with the appropriate visibility. The class acts as an object where it needs visibility into every other package and ever other package needs visibility to it. This creates a circular dependencies.

Additionally, it's more convenient for external usage of an API to resemble internal usage of the same API, so that upstreaming research code is easier.

PiperOrigin-RevId: 448995317