Rename up `compiler/test_utils` to `compiler/building_block_test_utils`.
This change is a follow up to cl/448978762, which organized the test utilities related to computations.

This change removes the need to alias the module when it is imported and makes it easier to reason about which test utilities we are using in the module. Additionally, this change applies this pattern more consistently across TFF.

PiperOrigin-RevId: 449352318