Disable test that were added using disallowed dependencies.
These tests rely on execution, more meaningfully they rely on code that the thing they are testing does not. This is possible because this visibility was added by mistake and has since been reverted.

This is problematic because it leads to having too much e2e coverage without having unit test coverage and because it make maintaining visibility/package structures difficult.

There are a few options to enable these tests:

1. Refactor the tests to not rely on execution
1. Move the tests to the appropriate place
1. Delete the tests

PiperOrigin-RevId: 451240151