Fixes stale state in C++ executor service refcounting in the case that a controller goes down without a chance to run its destructor.
The current behavior manifests in a subtle bug, with an artifically high refcount--the runtime cannot recover from a (sufficiently bad) controller failure, followed by a worker failure.

This change inspects the statuses in the executor service--if an executor underneath the service returns a FailedPrecondition status (therefore indicating it must have some state set before it can be used), outstanding references to this executor are by definition invalid. Therefore the service releases its handle on this executor, resetting its refcount to 0.

Verified to fix a reproduction of the issue. Leaves the issue inplace in the Python runtime, for now--invariants are less clear on the Python side.

PiperOrigin-RevId: 452410075