Allow for concurrent invocations of different cardinality computations in TFF async execution contexts.
The most exciting change here is the deletion of the ensure_closed parameter. This parameter protected stacks that failed in a partially-constructed state; remote workers could be left from this process with their cardinalities 'set'. Since the move to remote executor construction in the stack function, this parameter has not been necessary. The global nature of this 'reset' is now very difficult to reason about in a world in which servers can host multiple concurrent invocations, so it is natural to delete in this CL.

PiperOrigin-RevId: 455184092