Add a `keras_model_from_functional_weights` method to assign FunctionalModel weights back to a a Keras model.
This is intended to be used as the reverse of
`tff.learning.models.functional_model_from_keras`.

Note: this is usable only when the Keras model is being constructed in a graph
context (or inside a `tff.tf_computation`), as Keras creates different model
variables in eager vs graph contexts.

Addresses #3087 for graph context (and `tff.tf_computation`) usages>

PiperOrigin-RevId: 462642528