Remove `initialize` from MapReduceForm and introduce `type_signature` component.
This CL moves MapReduceForm from being 1:1 with an IterativeProcess to being
1:1 with a Computation.

We want to continue to be able to evaluate if computations are compatible
conceptually with single federated rounds for platforms that require
MapReduceForm. However, we also need our data structures to acknowledge that
some federated executions require different pairings of computations (2 or
more). Previously, MapReduceForm had components that aligned to both
IterativeProcess.initialize, namely the MapReduceForm's `initialize` component,
and IterativeProcess.next (everything else in the MapReduceForm). By pulling
the "initialization" part out of the MapReduceForm, we can align the
MapReduceForm exactly to a single Computation and have more flexibility when
defining the executions we want to run.

This CL removes get_map_reduce_form_from_iterative_process and replaces it with
get_map_reduce_form_from_computation and
get_state_initialization_computation_for_map_reduce_form. The former returns a
MapReduceForm from a tff.Computation; the latter returns a tff.Computation
containing just a TensorFlow function, identical to the removed `initialize`
component of MapReduceForm.

The `initialize` component of MapReduceForm was also used to verify that the
type signatures of interconnected components matched up and also helped to
ensure that inputs would have the correct type signature for IterativeProcesses
in MapReduceForm. Since we removed `initialize`, we introduced added a
`type_signature` attribute of MapReduceForm. When constructing an instance of
MapReduceForm, the type signature of the Computation used to generate
MapReduceForm needs to be passed in as an argument.

Additionally, the testing examples in mapreduce_test_utils.py were updated to
return both an initialization Computation and a MapReduceForm.

PiperOrigin-RevId: 466361463