Remove the downgrade to `pip`.
This was introduced in cl/466997603. The compatibility specifier issues is still present in this version of `pip`. Rather than downgrading `pip`, we will stop using the compatibility specifier in the cases where the version ends in "0" and use the expanded form instead. For exmaple, use `>=x.y,==x.*` instead of `~=x.y`.

PiperOrigin-RevId: 468279815