Adds support for SparseTensor-yielding datasets in remote and C++ runtimes.
Generally, this CL takes the approach of forcing datasets which are intended to yield SparseTensors to rather yield OrderedDicts with the appropriate three-element SparseTensor representation. This approach was chosen because direct iteration over SparseTensor-yielding datasets in C++ yields a *variant* tensor (which can be deserialized into the three-element representation we expect); this behavior does not match the way TFF prefers to represent SparseTensors (and generally composite tensors), and so would require a (conditional) runtime conversion if we did not perform this 'eager structuring'. With some thought, I preferred to take this approach for consistency and relative ease.

I've tried to keep the implicit couplings as small as possible, and call them out explicitly where they still happen. They can, I think, all be understood by the mental model: inside of TFF we represent SparseTensors as odicts with particular keys. It is the job of computation serialization, as well as argument ingestion and result transformation (on the way back to the user) to ensure that the user's expectations are matched, with this internal representation as 'the contract' respected by the software.

PiperOrigin-RevId: 491339551