Remove default value for the key parameter from the abstract base class `tff.program.ReleaseManager`.
The original intention of this API was for key to be optional. However, instead of making it optional I gave it a default value. In some subclasses you cannot provide a reasonable default value for the key, therefore it's not appropriate for the abstract base class to provide a default value. For now we can remove the default value. I have cl/473018138, which should enable program logic to make this parameter optional.

PiperOrigin-RevId: 492835929