Enable and fix python annotations in `execution_contexts` package.
The `_create_local_python_execution_context` function had an awkward type annotation that returned a `Union` but callers would always return one of the elements of the Union and the linter was complaining. This could have been fixed using Generic and TypeVar, but fortunately it was just simpler to delete the function.

PiperOrigin-RevId: 500250973