[XLA:Python] Change Tensorflow federated XLA backend to pass MHLO to compile().
Change in preparation for dropping support for classic HLO arguments to compile().

This change has two parts:
a) change the calling convention to avoid wrapping HLO parameters in a tuple. This causes problems with the MHLO conversion. The tuple appears unnecessary and it's simplest to just remove it.
b) add a classic HLO to MHLO conversion on the input to compile().

PiperOrigin-RevId: 501908211