Add the threadpool to the ComposingExecutor.
This does two things:
- Limits thread creation to a sized pool, scheduling work in a FIFO order on
  the pool.
- Forces a cleaner destruction of ComposingExecutor, by way of destructing the
  ThreadPool member which closes the pool to new work and then joins the
  threads in the pool.

The second of which removes the flakiness seen on Dispose call expectations in
unittests.

This did require removing the `keepalive_this = shared_from_this()` capture in
the CreateCall thread creation as this would have created a resource deadlock
during destruction: the new ThreadPool could have threads which a shared_ptr to
the ComposingExecutor attempt preventing destruction. This is safe because the
threadpool ensures threads (that may have `this` captured) are joined, ensuring
they finish before the executor is destructed.

PiperOrigin-RevId: 511223019