Fix the formulation of the JIT constructed mapped selection computation that is sent to the remote machine in streaming mode.
Previously the `CreateSelectionFederatedStructComputation` function created one
federated_map that called a single selection function, this is not correct as
it effectively was creating an identity function.

To "push the placements to the leaves of the structure" we require a structure
of multiple `federated_map` calls, each doing performing one leaf selection.

PiperOrigin-RevId: 512084249