Fix control edge insertion logic for initialization ops.
We need to ensure that all the nodes that are requirements for 'group_initialization' have a control dependency injected correctly. We need to check for which graph nodes have same inputs as all nodes in group_initialization and insert a control edge. Previously check was incomplete, that led to missing some assign variable ops in the graph.

PiperOrigin-RevId: 518709294