Avoid placing iterator ops on GPUs.
We've seen errors in multi-GPU settings where iterator ops placed on GPUs
attempt to copy dataset/iterator variants to the GPU, which is not supported,
causing errors to be raised. Likely TFF's explicit device logic is interfering
with TF device placement logic at runtime. However it was not sufficient to
simply skip setting `device` attributes on these ops, it was required to
explicitly set them to the CPU.

This was found using FilePerUserClientData based datasets, for example the
FLAIR implementation.

PiperOrigin-RevId: 522655638