Update `_Intern` metaclass to be more friendly with Python type analyzers.
* Updated the implementation of `_Intern`.
* Standardized the args for `StructType.__init__`
* Added tests to ensure that the paramters to `_hashable_init_args` and `__init__` for a given interned `cls` remained the same.
* Fixed some unit tests.

Note: this change should largely be a noop. There is one important difference:

Previously the lifecycle for interning looked like:

```
1. _Intern.__call__
2. normlaized_args = cls._normalize_init_args(original_args)
3. ... = cls._hash_normalized_args(normlaized_args)
4. ... = super().__call__(normlaized_args)
(which then calls)
5. cls.__init__(normlaized_args)
```

This change updates this lifecycle to look like:

```
1. _Intern.__call__
3. ... = cls._hashable_init_args(original_args)
4. ... = super().__call__(original_args)
(which then calls)
5. cls.__init__(original_args)
```

The impact is that args being passed to `_Intern.__call__` and to `cls.__init__` are not being mutated, and therefore Python type analyzers can validate the Python types passed into the initializer correctly.

This also makes it easier for humans to read this code. For example, `StructType.__init__` took a different set of args than `StructType._normalize_init_args`, which made it very confusing to drive by and make changes without internalizing what was happening with the interning code.

For testing:
* pytype is still disabled, I am testing manually that these changes will allow pytype to be enabled for this target by patching this change into cl/529794709 and building running pytype on the `computations_types` module.
* I am relying on the existing unit test and benchmarks to ensure that the refactor to `_Intern` did not cause regressions.
* I have searched for code that called `StructType.__init__` with an unnamed "convert" to find errors with the API change and I was unable to find any.

PiperOrigin-RevId: 530985240