Remove compression for worker_binary.
Decompressing the binary on first usage causes problems because the target
location location may not be writable (e.g., the wheel was installed as a
different user or the program is running in a sandbox). It also introduces
a race condition if two processes load TFF at the same time (e.g., if bazel
runs multiple tests concurrently).

This increases the package size from 39 MB to 65 MB.

PiperOrigin-RevId: 533253036