Replace the `tff.framework.ComputationBuildingBlock.is_foo` methods with `isinstance`.
This change is a follow up to cl/510264449 (which fixed the pytype errors caused by enabling pytype in the `building_blocks` module) and removes the methods causing many of these issues.

This change makes it easier to enable pytype for the `computation_types` module because if the type of the `ComputationBuildingBlock` is known then the `type_signature` of the  `ComputationBuildingBlock` may be known as well. For example, a `tff.framework.Struct` will have a `type_signature` equal to `tff.StructType`

PiperOrigin-RevId: 540681787