Small bugfix in TFF SGDM optimizer.
Previously, initializing with 0.0 momentum caused _MOMENTUM_KEY to be tracked in the optimizer's hparams, but not be tracked in the optimizer's state. This caused the implementation of get_hparams to fail with a missing key. We could imagine tracking momentum 0 in the state; but if so, we would want to also track the accumulator (though it should be constant 0), or alter the `next` logic to case on the truthiness of the momentum value rather than presence or absence of the key.

The choice here seemed to me closest to the original intention as I read the code, so went with it. But happy to revise.

PiperOrigin-RevId: 548216337