Fix Python 3.11 related errors.
Every time we test a new version of Python the tests that assert against the traceback fail. Because TFF is not a Python library for writing error messages the traceback should be considered an implementation detail and should not be asserted against.

This change fixes these tests to assert that the expected error is thrown and removes the golden API that enables these kinds of tests.

Note: `self.assertRaisesRegex` has the same issue. If you find yourself needing to distinguish between one error and another, it's often better to create a custom error message than to assert against the error of the message.
PiperOrigin-RevId: 567620704