Remove usage of `Any` from modules in the `function_utils` module.
In most cases it's possible to simply replace `Any` with `object` but in some cases a more specific type is appropriate.

This change also fixes any pytype errors caused by adding the correct Python type annotation.

Note: _Arguments is deleted to make the code more readable, it's not a first class object and there is no parity between the packing inputs and the unapcking outputs.
PiperOrigin-RevId: 568246032