Bug fix: skip searching for `kServerUri` in the cardinality map when materializing values in the streaming remote exeuctor.
The `GetExecutor` call does not populate the server placement, as it infers
from the CLIENTS placed incoming value, and the stack's assumption is that it
is always `1`.

Also preemptively tries to check if the gRPC call will fail due to message
serialization and returning an error instead of CHECK failing.

PiperOrigin-RevId: 582083907