Remove `tf.RaggedTensor` and `tf.SparseTensor` from the `federated_sum`'s checks.
These checks were extended in cl/367261866 with the following note:

```
In order to prevent accidental summation of the dense representation of RaggedTensor and SparseTensor, `federated_sum`'s checks are extended to ensure that no `RaggedTensor` or `SparseTensor` container types are present. However, this check does not extend without effort to other non-intrinsic aggregations, so ensuring that the dense representations of RaggedTensor and SparseTensor are not natively summed is left to library authors and end-users.
```

However, this change introduces a TensorFlow dependency in the federated context. Instead it is the responsibility of the executors to determine if these types of values can be summed or the responsibility of the library authors and end-users to ensure that these values are not passed to the intrinsics in the first place.

PiperOrigin-RevId: 596757008