Remove usages of `infer_type`.
In order to facilitate this, we add a (non-public) method that creates a `tff.StructType` corresponding to a `ReconstructionModel`'s global variables. This method uses `tf.nest.map_structure` in order to map model variables to TFF types, and relies on the fact that we know that leaf nodes are `tf.Variable`s.

PiperOrigin-RevId: 598847925