Remove a `TensorFlow` dependency from the `computation_wrapper` module.
Update the decorators to case on `Callable` types rather than try and define a set of callables to allow.

TFF supports passing type specifications (i.e. objects that can be turned into a `tff.Type`) as arguments to a computation decorator. In some cases those type specifications (e.g. np.int32) are a `type`, making them `Callable`, but they should not be treated as the function being decorated.

PiperOrigin-RevId: 602505595