Remove the runtime-agnostic support for `tf.RaggedTensor` and `tf.SparseTensor`.
These are TensorFlow-specific constructs and **should** be usable within a TensorFlow computation but non-TensorFlow computations **should not** have to handle these constructs. Currently, other kinds of TFF comptuations (but not all) are handling some (but not all) TensorFlow constructs. This is preventing those environments from being usable without a TensorFlow dependency and it's cuasing inconsistency in how different kinds of computations can be used together.

For example, consider the following JAX computation:

```
@tff.jax.computation()
def foo:
  ...
```

* The person who writes the JAX computation doesn't expect to have to install non-JAX dependencies (e.g. TensorFlow) in their environment.
* The person who writes the framework for creating JAX computation doesn't expect to have to handle non-JAX constructs (e.g. TensorFlow objects).

This change should make it easier for users to use TFF and to reason about where the boundary of a specific environment exists within a federated program.

PiperOrigin-RevId: 615977435