Update new `content` field to be an `Any` proto instead of bytes.
Why Any proto:
Using an Any proto forces the field to be a proto meaning we get the clear backwards compatibility rules and checks of protos.

Argument for bytes:
Using bytes gives us opacity of type, making it clear that this is a contract between a platform and itself, meaning other parts of the TFF execution stack should not inspect/modify it - only platform-specific executors.

We choose to make this an Any proto for the backwards compatibility benefits.

PiperOrigin-RevId: 626391456