Remove unnecessary (somewhat buggy) golden test case check from `compiler_test_utils_test`.
This test case uses the same file as a different test case. When generating an updated golden test file with --test_arg=--update_goldens --test_strategy=local this will actually break this test case because a correct file is generated for the test, not allowing the test to fail.

We decide to remove this test case as it is unneeded.

PiperOrigin-RevId: 629171693