Update the `WORKSPACE` file to use `http_archive` rules instead of `git_repository` rules to install dependencies.
The `git_repository` was used previously because it is easier to use and it is more clear what the dependency is pinned to (i.e. branch, commit, tag).

The `http_archive` has some advantages and hopefully with documentation we can still provide guidance on how to depend on released products (i.e. tags). The primary advantage of the `http_archive` rule is that it works with https://bazel.build/run/build#repository-cache.

PiperOrigin-RevId: 633711203