Add `__iter__` method to `tff.StructType`.
This overrides the `tff.structure.Struct` implementation that yields values with an implementation that yields name/value pairs. And is useful to reduce the need for `tff.structure.iter_elements` and to deprecate usage of `tff.structure.*` API altogether.

This change effectively moves `tff.StructType` from iterating like a tuple and to iterating like a named tuple.

PiperOrigin-RevId: 645167339