Save an immutable data source iterator object in the program state.
The data source iterator is saved in the program state to allow the program to use the same data selection configuration when resuming from a previous round. If the iterator is a mutable object, when the async `program_state_manager.save` is called, the iterator might have been updated for the next round, and the incorrect information would be saved in the program state for the current round and be used for future rounds after resuming. Making the saved iterator an immutable object can help solve this problem.

PiperOrigin-RevId: 650681494