Update the implementation of `array.is_compatible_dtype`.
This function attempted to apply value-based logic for `np.ndarray` types. However, this was causing performance issues with very large `np.ndarray`s and often resulted in timeouts.

This change aligns how TFF and numpy test if an `np.ndarray` can be cast to a particular dtype and add some more testing and documentation.

PiperOrigin-RevId: 658460040