Get rid of CompositeKeyOrderedVector in the CompositeKeyCombiner
Instead of maintaining CompositeKeyOrderedVector, which is a vector of
pointers to composite keys, we can simply create this vector from the
map of composite keys when we need it at the very end. That can be done in
one pass over the map. In addition to not having to maintain and grow the
vector, we can also move from node_hash_map to flat_hash_map, since we no
longer need to maintain the pointer stability. flat_hash_map should give us a more compact in-memory representation of composite keys.

PiperOrigin-RevId: 684197137