Composite Key Combiner optimization.
This changes Composite key implementation from FixedArray to std::string - the way it was implemented originally. The reason it was changed to FixedArray is to deal with unaligned pointers. This implementation uses memcpy to save and load data to/from std::string, so unaligned memory access is no longer a concern. I've verified that the compiler replaces memcpy calls with inlined intrinsic since the number of copied bytes is known at compile time - it is either 4 bytes or 8 bytes. Copying of values to/from composite keys is as optimal as it can be and memcpy doesn't show up in the CPU profile.

Benchmarking shows ~45% improvement in throughput of inserting composite keys when combining 16 keys because hashing of strings is more efficient. There is also 20-40% reduction in peak memory use depending on whether the key values are 32 bit or 64 bit.

PiperOrigin-RevId: 686560637