Properly handle empty `Tensor`s in `AggVectorIterator`.
When a `Tensor`'s `TensorData` has a non-null but empty `data` buffer, then before
this CL `t->AsAggVector().begin() == t->AsAggVector().end()` was not true, which is
a bug and results in non-existent elements being accessed in the vector.

This change fixes that issue by detecting when an empty `TensorData` is passed to the
`AggVectorIterator` constructor, and setting the `ptr` and `end_ptr` values to
nullptr in that case, which ensures that the created iterator equals the `end()`
iterator.

PiperOrigin-RevId: 702136639