Refactor the sequence executor to avoid relying on TensorFlow.
Use an array of Arrays instead of a TF dataset to represent a deserialized sequence in the sequence executor. Note the serialized representation of a sequence is already an array of Arrays. Meaning that the sequence executor was converting the Array proto into a dataset and then back into an Value proto that contained the Array proto.

PiperOrigin-RevId: 738926912