Add Tensor constructors for scalar tensors and one-dimensional tensors.
This adds an ability to construct a tensor from a numeric or string scalar value, or from a vector, or from an initializer_list. The type of the tensor and the shape is deducted from the argument. Please note that construction will always succeed in these cases except when the type is unsupported, but since the type is known at compile time there is no need to return StatusOr. The more general construction method via Tensor::Create is still available for the cases when the simple construction is insufficient. But using these new constructors will make the code more concise and readable.

Examples:

// Scalar tensors
Tensor t1(123);
Tensor t2(123.4f);
Tensor t3("abc");  // string_view or string&, or string&& are accepted too

// One dimensional tensors
Tensor t4({1, 2, 3}); // std::vector<T>&& is accepted too
Tensor t5({"a", "b", "c"}); // std::vector<std::string>&& is accepted too

PiperOrigin-RevId: 840782940